modelAPI/requirements.txt
anyio==4.6.0
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Cors==5.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.1.1
ollama==0.3.3
openpyxl==3.1.5
pandas==2.2.3
pymongo==4.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
tzdata==2024.2
Werkzeug==3.0.4
wheel==0.44.0
